Rationalise some code
